crawler/requirements.txt
beautifulsoup4==4.12.3
requests==2.31.0
lxml==5.1.0
python-dotenv==1.0.1
supabase==2.3.4
pyyaml==6.0.1
